Forgot one thing

+++ src/CRE/render_context.cpp
@@ -1770,7 +1770,7 @@ namespace mdl {
         if (!buff)
             return 0;
 
-        int32_t size = align_val(sizeof(ObjKind) + sizeof(mat4) + sizeof(float_t) * 2, 0x08);
+        int32_t size = (int32_t)align_val(sizeof(ObjKind) + sizeof(mat4) + sizeof(float_t) * 2, 0x08);
         switch (kind) {
         case OBJ_KIND_NORMAL:
             size += sizeof(ObjSubMeshArgs);
